Move function and types for decoding readings to new module.
Also added some doc comments.
